requirements.txt
langchain-community
python-multipart
openai
fastapi
tiktoken
accelerate
bitsandbytes
faiss_cpu
ctransformers
sentence_transformers
unstructured
python-dotenv
langchain-openai
crewai
crewai[tools]
google-cloud-storage
cloud-sql-python-connector==1.9.2
fastapi==0.111.0
google-auth==2.29.0
pg8000==1.31.2
SQLAlchemy==2.0.30
passlib==1.7.4
python-jose==3.3.0
langsmith